requirements.txt
django==2.2.10
djangorestframework
psycopg2